# Conflicts:
#   frontend/src/store/index.js
#   frontend/src/views/HomeView.vue

+++ backend/travel-on/src/main/java/com/travel/travel_on/model/service/NoticeServiceImpl.java
@@ -1,7 +1,7 @@
 package com.travel.travel_on.model.service;
 
-import com.travel.travel_on.dto.FAQ;
-import com.travel.travel_on.dto.Notice;
+import com.travel.travel_on.entity.FAQ;
+import com.travel.travel_on.entity.Notice;
 import com.travel.travel_on.model.repo.FAQRepository;
 import com.travel.travel_on.model.repo.NoticeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,28 +19,18 @@ public class NoticeServiceImpl implements NoticeService {
     @Autowired
     FAQRepository fRepo;
 
-//    @Override
-//    public List<Notice> select() {
-//        List<Notice> result = nRepo.findAll();
-//        if (!result.isEmpty()){
-//            return result;
-//        }
-//        return null;
-//    }
-
     @Override
     public Page<Notice> findPage(Pageable pageable) {
         return nRepo.findAll(pageable);
     }
 
     @Override
-    public int write(Notice notice) {
-        Optional<Notice> result = nRepo.findById(notice.getNoticeId());
-        if (result.isPresent()){
-            return 1;
+    public boolean write(Notice notice) {
+        if (notice == null){
+            return false;
         }else {
             nRepo.save(notice);
-            return 0;
+            return true;
         }
     }
 
@@ -55,38 +45,32 @@ public class NoticeServiceImpl implements NoticeService {
     }
 
     @Override
-    public int update(Notice notice) {
+    public boolean update(Notice notice) {
         Optional<Notice> result = nRepo.findById(notice.getNoticeId());
         if (result.isPresent()){
             nRepo.save(notice);
-            return 0;
+            return true;
         }else{
-            return 1;
+            return false;
         }
     }
 
     @Override
-    public int delete(Integer id) {
+    public boolean delete(Integer id) {
         Optional<Notice> result = nRepo.findById(id);
-        result.ifPresent(notice -> {
-            nRepo.delete(notice);
-        });
-        return 1;
+        if(result == null){
+            return  false;
+        }else{
+            result.ifPresent(notice -> {
+                nRepo.delete(notice);
+            });
+            return true;
+        }
     }
-
-//    @Override
-//    public List<FAQ> selectFAQ() {
-//        List<FAQ> result = fRepo.findAll();
-//        if (!result.isEmpty()){
-//            return result;
-//        }
-//        return null;
-//    }
 
     @Override
     public Page<FAQ> search(String keyword, Pageable pageable) {
         return fRepo.findByQuestionContaining(keyword, pageable);
     }
-
 
 }
